move UserRepository
